\name{categoricalexample}
\alias{categoricalexample}

\title{
    A small dataset for showing how partialNA works in freqparcoord.cd with
    categorical variables
}

\description{
    A small fictional dataset with different values and NA's to emphasize
    partialNA and frequency based calculations with freqparcoord.cd
    and categorical variables.
}
\usage{
data(categoricalexample); categoricalexample
}
