\name{findParentTuples}
\alias{findParentTuples}

\title{
    Given a row and a list of lists of intact rows, find all possible substitutions for NA values.
}

\description{
}

\usage{
    findParentTuples(row, intact)
}

\arguments{
    \item{row}{
        The row with NA values.
    }
    \item{intact}{
        The intact dataset to fill values from.
    }
}

\details{
}

\value{
    Object of type "matrix".
}

\author{
Norm Matloff <matloff@cs.ucdavis.edu>, Harrison Nguyen <hhnguy@ucdavis.edu>, and Vincent Yang <vinyang@ucdavis.edu>
}

\examples{
}
