\name{freqparcoord.cd}
\alias{freqparcoord.cd}

\title{
    Parallel coordinates.
}

\description{
}

\usage{

}

\arguments{

}

\details{

}

\value{

}

\author{
Norm Matloff <matloff@cs.ucdavis.edu>, Harrison Nguyen <hhnguy@ucdavis.edu>, and Vincent Yang <vinyang@ucdavis.edu>
}

\examples{
data(mlb)
}
