\name{hrdata}
\alias{hrdata}

\title{
    A human resources simulated dataset.
}

\description{
    A small fictional dataset by Kaggle that includes satisfaction level,
    the result of their last evaluation, number of projects, average monthly hours,
    time spent at the company, whether they have had a work accident, whether
    they have had a promotion in the last 5 years, their department,
    salary and finally whether the employee has left the company. Each
    row represents a single employee.
}
\author{
    Norm Matloff <matloff@cs.ucdavis.edu>, Vincent Yang <vinyang@ucdavis.edu>, and Harrison Nguyen <hhnguy@ucdavis.edu>
}
\usage{
    data(hrdata); hrdata
}
