\name{reOrder} 
\alias{reOrder}    

\title{
    Wrapper for discretize.
}

\description{
    Use to order the levels of a factor in a desired sequence.
}

\usage{
    reOrder(dataset, colName, levelNames)
}

\arguments{
    \item{dataset}{
        Dataset to reorder.
    }
    \item{colName}{
        Column names to use.
    }
    \item{levelNames}{
        The names of the levels to use.
    }
}

\author{
    Norm Matloff <matloff@cs.ucdavis.edu>, Vincent Yang <vinyang@ucdavis.edu>, and Harrison Nguyen <hhnguy@ucdavis.edu>
}
