\name{showCounts} 
\alias{showCounts}    

\title{
    Show the tuple counts for the most recent operation.
}

\description{
    Use with saveCounts=TRUE to recover the tuple counts.
}

\usage{
    showCounts(nshow=NULL)
}

\arguments{
    \item{nshow}{
        The dataset to show.
    }
}

\author{
    Norm Matloff <matloff@cs.ucdavis.edu>, Vincent Yang <vinyang@ucdavis.edu>, and Harrison Nguyen <hhnguy@ucdavis.edu>
}
