\name{reOrder} 
\alias{reOrder}    

\title{
    Re-order levels of a factor, according to some desired ordinal form.
}

\description{
    Use to order the levels of a factor in a desired sequence.
}

\usage{
    reOrder(dataset, colName, levelNames)
}

\arguments{
    \item{dataset}{
        Dataset to reorder.
    }
    \item{colName}{
        Column name.
    }
    \item{levelNames}{
        Names of the reordered levels
    }
}

\author{
    Norm Matloff <matloff@cs.ucdavis.edu>, Vincent Yang <vinyang@ucdavis.edu>, and Harrison Nguyen <hhnguy@ucdavis.edu>
}
