\name{smallexample}
\alias{smallexample}

\title{
    A small dataset for showing how partialNA works in freqparcoord.cd
}

\description{
    A small fictional dataset with different values and NA's to emphasize
    partialNA and frequency based calculations with freqparcoord.cd.
}
\usage{
data(smallexample); smallexample
}
