\name{smallexample}
\alias{smallexample}
\alias{categoricalexample}

\title{
    A small dataset for showing how tupleFreqs works in cdparcoord
}

\description{
    A small fictional dataset with different values and NA's to emphasize
    tupleFreqs and frequency based calculations with cdparcoord.
}
\author{
    Norm Matloff <matloff@cs.ucdavis.edu>, Vincent Yang <vinyang@ucdavis.edu>, and Harrison Nguyen <hhnguy@ucdavis.edu>
}

