\name{smallexample}
\alias{smallexample}

\title{
    A small example demonstrating freqparcoord.cd.
}

\description{
    A small example showing how the freqparcoord.cd package creates
    parallel coordinate plots.
}

\usage{
smallexample(n)
}

\arguments{
    \item{n}{
        The optional number of top frequencies you'd like to select.
    }
}

\author{
Norm Matloff <matloff@cs.ucdavis.edu>, Harrison Nguyen <hhnguy@ucdavis.edu>, and Vincent Yang <vinyang@ucdavis.edu>
}

\examples{
    # This will draw a plot for the smallexample.csv dataset
    smallexample()
}
